requirements.txt
fastapi
uvicorn[standard]
kafka-python==2.0.2
sqlalchemy
psycopg2-binary==2.9.9
python-dotenv==1.0.0
pydantic==2.5.2
requests==2.31.0 
